requirements.txt
pyTelegramBotAPI
pymorphy2
pyyaml
psycopg2-binary
peewee
python-dotenv
sanic